Pass environment variables with empty values
There is a difference between an empty value and an unset environment
variable. We should not confuse both; If the user wants to unset an
environment variable, they can certainly do so (unsetenv(3), or in the
shell: 'unset ABC').

This fixes Git's t3301-notes.sh and t6006-rev-list-format.sh, which
override environment variables with empty values.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>