Tell non-MSys2 executables whether they are connected to a terminal
By setting the MSYS_TTY_HANDLES environment variable to a list of
Win32 handles corresponding to terminals connected to
stdin/stdout/stderr, we help non-MSys2 executables (which cannot simply
use MSys2's isatty()) to tell whether their standard file descriptors
are redirected or not.

To make things as convenient as possible, the list is delimited by
spaces on both sides of the elements so that to find out whether stdin
is actually a terminal, one just needs to look for the needle
" <handle> " in $MSYS_TTY_HANDLES (where <handle> is the decimal
representation of GetStdHandle(STD_INPUT_HANDLE)).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>