Introduce the 'usertemp' filesystem type
In the context of Git for Windows, the MSys2 root lives inside
%PROGRAMFILES% which is typically write-protected for mere mortals' user
accounts. In particular for /tmp/, this is a problem, because many
scripts assume that writing to /tmp/ is fair game.

MSys1 used to mount /tmp/ automatically, pointing to the current user's
temporary directory.

Let's recreate that functionality in a slightly different way: we now
interpret the filesystem type "usertemp" in the /etc/fstab file. To make
/tmp/ point to the temporary directory of the current user, as per the
%TMP% environment variable, just add this line to /etc/fstab:

	none /tmp usertemp binary,posix=0 0 0

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>