fixup! Allow overriding the home directory via the HOME variable
Oh wow. Cygwin's `sys_wcstombs()` has a subtle bug in that if you use
the 3-parameter version (i.e. passing an implicit (size_t)-1 as the
number of input characters), it will return the number of output bytes
*including* the NUL character (disagreeing with the documentation).

As a consequence, the previous way to expand $HOMEDRIVE$HOMEPATH into
a single buffer managed to expand *only* $HOMEDRIVE.

This fixes https://github.com/git-for-windows/git/issues/414 and
https://github.com/git-for-windows/git/issues/434.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>