fixup! Add functionality for converting UNIX paths in arguments and environment variables to Windows form for native Win32 applications.
There is a subtle buffer overrun in the convert() mechanism: after a
sub_convert() completed, we may very well be at the end of the
conversion, i.e. the "src" pointer may already point at the trailing
NUL. Advancing that pointer any further shifts the pointer out of
bounds, with the predictable result of copying bogus characters into the
output buffer.

Another very subtle buffer overrun happens when the the buffer passed to
the find_path_start_and_type() function ends in a colon: in that case,
the code advances one byte to see whether there is a slash, another
colon, or a period, or if there is no equal sign in th rest of the
buffer, but since there is only a NUL byte, the code enters the next
loop iteration, advancing by one byte *again*, and consequently failing
the `it2 != end` condition.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>