fixup??? Add functionality for converting UNIX paths in arguments and environment variables to Windows form for native Win32 applications.
The purpose of the subp_convert() function is to convert POSIX path lists'
individual components. As a special case, the code allows URLs to be left
alone.

However, the code in ppl_convert() to identify what constitutes is buggy,
as it is too loose in what it interprets as URL (the mere substring "://"
will fool it into believing that, say, "/c/://share/c$" constitutes a
URL), and when subp_convert() then asks the find_path_start_and_type()
function to determine the type, it correctly identifies this as a POSIX
path list, and then subp_convert() calls ppl_convert() again, causing an
infinite recursion. The symptom is a "Bad address" error message.

As a stop-gap measure (leaving the fix that teaches ppl_convert() to be
more careful in what it identifies as URLs for later), let's just force
subp_convert() to trust the caller's judgement when the `is_url` parameter
is non-zero.

This fixes https://github.com/git-for-windows/git/issues/1033

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>