fixup! Do not add a trailing slash when converting / to a Win32 path
After long discussions, and several attempts to address this on Git's
side (t0060-path-utils failed specifically when this patch was
reverted), finally Git is ready to have this patch dropped.

This is a step in the right direction: it reduces the difference between
the official MSYS2 runtime and Git for Windows' forked version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>