fixup! POSIX-ify the SHELL variable
Funnily enough, setting the SHELL variable to the empty string will be
mishandled without this fix when spawning a child process: the
build_env() method insists on using a temporary win_env for the
conversion that needs to be copied from the convvars (and hence wants
to copy the skip_if_empty bit, too).

It is unclear to this developer how this worked in 2.6.1-4 (it should
not have worked, but it clearly did, otherwise Git's test suite would
have failed in t3404-interactive-rebase's exec test which sets SHELL in
that very way), but it is an obviously correct bug fix in any case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>