Avoid `assert (this);` statements
It may have been legal C++ at some stage to run methods with `this`
pointing to `NULL`, but it is no longer legal to do so. For `assert`
statements, G++ v6.3.0 does not yet spit out errors, but mere warnings:

	warning: nonnull argument 'this' compared to NULL [-Wnonnull-compare]

It is only a matter of time until these warnings will be turned into
errors, so let's just get rid of those statements.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>