Refuse to send job control signals to Win32 processes
When calling Win32 processes from MSYS2, it does not really make sense
to kill them when a user tries to suspend them. It is much better to
warn the user that Win32 processes simply cannot be suspended.

This closes https://github.com/git-for-windows/git/issues/1083 and
replaces https://github.com/git-for-windows/build-extra/pull/154

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>