squash! Try to kill Win32 processes gently upon Ctrl+C
Let's redo this. We will need to squash this into the original Ctrl+C
patch during the next merging-rebase.

Handle SIGINT/SIGTERM specifically via Ctrl events

This assumes that we want to send the signal to a console process. It
would not make sense in any other case, anyway, to try to emulate a
Ctrl+C: where there is no Console, there are no keystrokes.

[expand explanation here]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>