Fix incorrect path conversion (trailing slash)
This commit reverts part of 8eae854a91 (Special case for converting root
directory to have training slash, 2016-03-01) which not only introduced
a funny typo, but also introduced overzealous behavior where the
path_conv class would keep trailing slashes in *waaaaaaay* too many
cases. Unfortunately, the original commit's message as well as the code
comment it introduced are seriously deficient in clarity and are only
excelling in their terseness that makes them pretty useless.

So the best we can do is assume that the change that caused the trailing
slashes to be kept in pretty much all cases (i.e. failing to normalize
the path `Y:\directory\` to `Y:\directory`) was introduced by mistake.

This mistake has serious repercussions, so the above rant has merit.
On FAT file systems, the normalized path is used to fake inode numbers.
As a result, `Y:\directory\` and `Y:\directory` have different inode
numbers!!!

One consequence of this bug has been reported on the Git for Windows:
the `find` and `rm` commands can error out on FAT file systems with very
confusing "No such file or directory" errors, for no good reason.

Vasil Minkov even pointed out that this bug had been fixed in Cygwin as
early as 1997... and we reintroduced it in MSYS2.

This fixes https://github.com/git-for-windows/git/issues/1497 and
https://github.com/git-for-windows/git/issues/1536.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>