squash! Fix incorrect path conversion (trailing slash)
The problem with the original approach is not that it is completely
bogus. After all, when you pass the option --prefix=/tmp/ to Git, you do
want to end up with a trailing slash.

The real problem with the original approach is that it simply changed
path_conv, completely oblivious and careless about other users of
path_conv.

That was really wrong, of course, and cost us time, sweat and tears. The
appropriate approach is to *not* affect other users, but instead
introduce a flag that we use in *our* caller, so that everybody gets
what they want:

- emulation of inodes on FAT by calculating the hash on the normalized
  path: works.

- MSYS2's conversion of "POSIX" paths to Windows paths: works.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>