fixup! Add functionality for changing OS name via MSYSTEM environment variables.
MSYS2 wants to add one extra byte to the `utsname` struct, to make sure
that `MINGW64` fits where `CYGWIN` would have been printed before.

In Cygwin v3.0.0, the `utsname` struct was changed, to allow for much
longer names. We also extended that when rebasing, but we failed to
re-extend the `old_utsname` which is there for previously-compiled
executables that do not know about the new `struct utsname`.

Let's change also `old_utsname`.

This fixes https://github.com/git-for-windows/git/issues/2081

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>