fixup! Pass environment variables with empty values
The patch was incomplete. It resulted in settings like `SHELL=` to
*skip* the variable name and the equal sign and instead insert a NUL,
i.e. instead of putting `SHELL=\0` into the environment block, it would
append `\0\0` which would be interpreted as the end of the environment
block instead.

Funnily enough, this was only noticed years later, when Git's `git
rebase -i` progress was changed to clear the line differently depending
on the `TERM` setting, and t/t3404-rebase-interactive.sh triggered the
bug by setting `SHELL` to the empty value, which due to the bug fixed
here would drop `TERM=dumb` by mistake.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>